fixes based on comments #5876
